Split context.rs
+ Introduces context module
+ Seperate files within module for
  + block,
  + cdf_context,
  + frame,
  + partition,
  + superblock,
  + transform
